LPS-115615 Move PortalExecutor
