+++ modules/apps/data-engine/data-engine-service/src/main/java/com/liferay/data/engine/internal/petra/executor/DataEngineNativeObjectPortalExecutor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.data.engine.internal.nativeobject;
+package com.liferay.data.engine.internal.petra.executor;
 
 import com.liferay.petra.concurrent.NoticeableExecutorService;
 import com.liferay.petra.concurrent.ThreadPoolHandlerAdapter;
